Split install detection into per-OS probes
UpdateCommandForCurrentBinary now walks installProbes and a per-OS
fallback. Strings:

- brew upgrade --yes entire / entire@nightly: versioncheck_unix.go
- curl install.sh (+ nightly): versioncheck_unix.go fallback
- mise upgrade entire: versioncheck.go miseProbe
- scoop update / rename migration: versioncheck_windows.go
- install.ps1 stable/nightly: versioncheck_windows.go fallback

Entire-Checkpoint: 01M1NZ65NZ2XJDRB2JK7KDKMP6
