+++ cmd/entire/cli/versioncheck/autoupdate_test.go
@@ -55,7 +55,7 @@ func useBrewExecutable(t *testing.T) {
 	t.Helper()
 	orig := executablePath
 	executablePath = func() (string, error) {
-		return "/opt/homebrew/Cellar/entire/1.0.0/bin/entire", nil
+		return "/opt/homebrew/Caskroom/entire/1.0.0/entire", nil
 	}
 	t.Cleanup(func() { executablePath = orig })
 }
@@ -65,17 +65,7 @@ func useMiseExecutable(t *testing.T) {
 	t.Helper()
 	orig := executablePath
 	executablePath = func() (string, error) {
-		return "/home/user/.local/share/mise/installs/entire/1.0.0/bin/entire", nil
-	}
-	t.Cleanup(func() { executablePath = orig })
-}
-
-// useScoopExecutable points the install-manager detector at a scoop install path.
-func useScoopExecutable(t *testing.T) {
-	t.Helper()
-	orig := executablePath
-	executablePath = func() (string, error) {
-		return `C:\Users\test\scoop\apps\cli\current\entire.exe`, nil
+		return miseExecutablePath, nil
 	}
 	t.Cleanup(func() { executablePath = orig })
 }
@@ -173,130 +163,34 @@ func TestMaybeAutoUpdate_NonTerminalWriter(t *testing.T) {
 	assertManualHint(t, buf.String(), brewUpgradeCmd)
 }
 
-// TestMaybeAutoUpdate_WindowsUnknownInstallerNoAutoRun verifies that on
-// Windows without Scoop or mise we print the install.ps1 one-liner and
-// never auto-run it (a running entire.exe cannot replace itself). The
-// POSIX curl fallback must not appear.
-func TestMaybeAutoUpdate_WindowsUnknownInstallerNoAutoRun(t *testing.T) {
-	tests := []struct {
-		name           string
-		currentVersion string
-		wantCmd        string
-	}{
-		{
-			name:           "stable",
-			currentVersion: "1.0.0",
-			wantCmd:        windowsInstallCmd,
-		},
-		{
-			name:           "nightly",
-			currentVersion: "1.0.1-nightly.202604101200.abc1234",
-			wantCmd:        windowsInstallNightlyCmd,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			f := newAutoUpdateFixture(t)
-			orig := executablePath
-			executablePath = func() (string, error) {
-				return windowsProgramFilesPath, nil
-			}
-			t.Cleanup(func() { executablePath = orig })
-
-			origGOOS := goos
-			goos = goosWindows
-			t.Cleanup(func() { goos = origGOOS })
-
-			var buf bytes.Buffer
-			action := maybeAutoUpdate(context.Background(), &buf, tt.currentVersion, "v2.0.0")
-
-			if f.installCalls != 0 {
-				t.Errorf("installer was auto-run on Windows + unknown install manager")
-			}
-			if action != autoUpdateActionSkip {
-				t.Errorf("action = %q, want %q", action, autoUpdateActionSkip)
-			}
-			out := buf.String()
-			if !strings.Contains(out, "when entire is not running") {
-				t.Errorf("missing Windows manual-run hint: %q", out)
-			}
-			if !strings.Contains(out, "  "+tt.wantCmd) {
-				t.Errorf("manual hint missing command %q: %q", tt.wantCmd, out)
-			}
-			if strings.Contains(out, "curl -fsSL") {
-				t.Errorf("Windows fallback must not show POSIX curl command: %q", out)
-			}
-			if strings.Contains(out, "download the latest release") {
-				t.Errorf("Windows unknown installer must not point at the releases page: %q", out)
-			}
-		})
-	}
-}
-
 // TestMaybeAutoUpdate_WindowsNeverAutoRuns verifies that on Windows the update
 // is never auto-run — a running entire.exe can't replace its own shim — so the
-// command is printed for the user to run once entire has exited. This holds for
-// every auto-installable manager on Windows, not just Scoop: mise is covered so
-// narrowing the branch back to Scoop can't silently regress it.
+// command is printed for the user to run once entire has exited. mise is the
+// shared manager covered here so narrowing the branch back to Scoop can't
+// silently regress it; Scoop and install.ps1 live in autoupdate_windows_test.go.
 func TestMaybeAutoUpdate_WindowsNeverAutoRuns(t *testing.T) {
-	tests := []struct {
-		name    string
-		setup   func(*testing.T)
-		wantCmd string
-	}{
-		{
-			name:    "scoop cli app prints migration command",
-			setup:   useScoopExecutable,
-			wantCmd: `cmd.exe /D /C "scoop install entire/entire && scoop uninstall entire/cli && scoop reset entire"`,
-		},
-		{
-			name:    "mise prints upgrade command",
-			setup:   useMiseExecutable,
-			wantCmd: "mise upgrade entire",
-		},
-		{
-			name: "unknown prints install.ps1 command",
-			setup: func(t *testing.T) {
-				t.Helper()
-				orig := executablePath
-				executablePath = func() (string, error) {
-					return windowsLocalBinPath, nil
-				}
-				t.Cleanup(func() { executablePath = orig })
-			},
-			wantCmd: windowsInstallCmd,
-		},
+	f := newAutoUpdateFixture(t)
+	useMiseExecutable(t)
+
+	origGOOS := goos
+	goos = goosWindows
+	t.Cleanup(func() { goos = origGOOS })
+
+	var buf bytes.Buffer
+	action := maybeAutoUpdate(context.Background(), &buf, "1.0.0", "v2.0.0")
+
+	if f.installCalls != 0 {
+		t.Fatalf("installer must not auto-run on Windows; calls=%d", f.installCalls)
 	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			f := newAutoUpdateFixture(t)
-			tt.setup(t)
-
-			origGOOS := goos
-			goos = goosWindows
-			t.Cleanup(func() { goos = origGOOS })
-
-			var buf bytes.Buffer
-			action := maybeAutoUpdate(context.Background(), &buf, "1.0.0", "v2.0.0")
-
-			if f.installCalls != 0 {
-				t.Fatalf("installer must not auto-run on Windows; calls=%d", f.installCalls)
-			}
-			// Plain skip, not skip-until-next-version: Windows gets no prompt,
-			// so there is no per-version choice to persist.
-			if action != autoUpdateActionSkip {
-				t.Errorf("action = %q, want %q", action, autoUpdateActionSkip)
-			}
-			out := buf.String()
-			if !strings.Contains(out, "when entire is not running") {
-				t.Errorf("missing Windows manual-run hint: %q", out)
-			}
-			if !strings.Contains(out, "  "+tt.wantCmd) {
-				t.Errorf("manual hint missing command %q: %q", tt.wantCmd, out)
-			}
-		})
+	if action != autoUpdateActionSkip {
+		t.Errorf("action = %q, want %q", action, autoUpdateActionSkip)
+	}
+	out := buf.String()
+	if !strings.Contains(out, "when entire is not running") {
+		t.Errorf("missing Windows manual-run hint: %q", out)
+	}
+	if !strings.Contains(out, "  mise upgrade entire") {
+		t.Errorf("manual hint missing command %q: %q", "mise upgrade entire", out)
 	}
 }
