+++ cmd/entire/cli/versioncheck/versioncheck.go
@@ -32,7 +32,6 @@ const (
 	installManagerBrew    = "brew"
 	installManagerMise    = "mise"
 	installManagerScoop   = "scoop"
-	installManagerUnknown = "unknown"
 	installChannelStable  = "stable"
 	installChannelNightly = "nightly"
 )
@@ -367,114 +366,53 @@ func normalizedExecPath() (string, error) {
 	return strings.ReplaceAll(filepath.ToSlash(realPath), "\\", "/"), nil
 }
 
-// scoopAppName returns the Scoop app directory the running binary lives under
-// — the path segment after `/scoop/apps/` (e.g. "cli" or "entire") — or ""
-// when the binary is not a Scoop install. This is the durable signal for the
-// pre-rename package: a binary running from the `cli` app dir must migrate to
-// `entire` regardless of its version (the fix ships in a final `cli` release,
-// so the migrating binary's version is already past the rename).
-func scoopAppName() string {
-	normalized, err := normalizedExecPath()
-	if err != nil {
-		return ""
-	}
-	_, rest, ok := strings.Cut(normalized, "/scoop/apps/")
-	if !ok {
-		return ""
-	}
-	// Cut returns the whole string when the separator is absent, so this covers
-	// both `.../scoop/apps/cli/current/entire.exe` and a bare app segment.
-	app, _, _ := strings.Cut(rest, "/")
-	return app
-}
-
-func installManagerForCurrentBinary() string {
-	normalizedPath, err := normalizedExecPath()
-	if err != nil {
-		return installManagerUnknown
-	}
-
-	switch {
-	case strings.Contains(normalizedPath, "/Cellar/") ||
-		strings.Contains(normalizedPath, "/opt/homebrew/") ||
-		strings.Contains(normalizedPath, "/linuxbrew/") ||
-		strings.Contains(normalizedPath, "/Caskroom/"):
-		return installManagerBrew
-	case strings.Contains(normalizedPath, "/mise/installs/"):
-		return installManagerMise
-	case strings.Contains(normalizedPath, "/scoop/apps/"):
-		return installManagerScoop
-	default:
-		return installManagerUnknown
-	}
-}
-
 // downloadsURL is the GitHub releases page, used for release-notes links.
 const downloadsURL = "https://github.com/entireio/cli/releases"
 
-// Windows install.ps1 one-liners from the README. Printed, never auto-run.
-const (
-	windowsInstallCmd        = "irm https://raw.githubusercontent.com/entireio/cli/main/scripts/install.ps1 -UseBasicParsing | iex"
-	windowsInstallNightlyCmd = "& ([scriptblock]::Create((irm https://raw.githubusercontent.com/entireio/cli/main/scripts/install.ps1 -UseBasicParsing))) -Channel nightly"
-)
+// installProbe identifies one install manager from the running binary's path
+// and returns the command that upgrades it. roots is unused until relocated
+// install dirs are honored; matches is marker-only.
+type installProbe struct {
+	name    string
+	roots   func() []string
+	markers []string
+	command func(execPath, currentVersion string) string
+}
+
+func noInstallRoots() []string { return nil }
+
+func (p installProbe) matches(execPath string) bool {
+	for _, m := range p.markers {
+		if strings.Contains(execPath, m) {
+			return true
+		}
+	}
+	return false
+}
+
+func miseUpgradeCommand(_, _ string) string {
+	return "mise upgrade entire"
+}
+
+var miseProbe = installProbe{
+	name:    installManagerMise,
+	roots:   noInstallRoots,
+	markers: []string{"/mise/installs/"},
+	command: miseUpgradeCommand,
+}
 
 // UpdateCommandForCurrentBinary returns the shell command that updates this
 // binary, based on how it was installed.
 func UpdateCommandForCurrentBinary(currentVersion string) string {
-	switch installManagerForCurrentBinary() {
-	case installManagerBrew:
-		if releaseChannel(currentVersion) == installChannelNightly {
-			return "brew upgrade --yes entire@nightly"
+	path, err := normalizedExecPath()
+	if err == nil {
+		for _, p := range installProbes {
+			if p.matches(path) {
+				return p.command(path, currentVersion)
+			}
 		}
-		return "brew upgrade --yes entire"
-	case installManagerMise:
-		return "mise upgrade entire"
-	case installManagerScoop:
-		// The Scoop package was renamed from `cli` to `entire`. A binary still
-		// running from the old `cli` app dir can never cross the rename with a
-		// plain `scoop update entire/cli`, so migrate it: install the new
-		// `entire` package, remove the old `cli` package, then reset the shared
-		// shims. Both manifests declare `bin: [git-remote-entire.exe,
-		// entire.exe]`, so the two packages contend for the same shims. Current
-		// Scoop handles that itself: installing `entire` renames the displaced
-		// shim to `entire.shim.cli` (warn_on_overwrite) and uninstalling `cli`
-		// removes only that suffixed copy (rm_shim), leaving the active shims
-		// pointing at `entire`. The reset step is belt-and-braces for older
-		// Scoop versions that overwrote shims without that alt-file dance and
-		// would otherwise take `entire.exe` and the git remote helper with them.
-		//
-		// The Windows updater is print-only, so return a self-contained command
-		// users can paste into either cmd or PowerShell. Explicitly invoking
-		// cmd.exe makes `&&` portable across those shells and ensures uninstall
-		// only runs after install succeeds — a stale bucket or transient failure
-		// therefore never removes the working CLI. (The uninstall→reset link is
-		// weaker: `scoop uninstall` ends in an unconditional `exit 0`, so reset
-		// runs even when the uninstall skipped its work. That is harmless, since
-		// reset is a no-op on the Scoop versions that make it unnecessary.)
-		//
-		// `scoop install` also auto-refreshes the bucket when >3h stale
-		// (is_scoop_outdated), so the new manifest usually lands
-		// without an explicit refresh; a bucket clone older than that check can
-		// still fail with "couldn't find manifest", and the README migration
-		// section names `scoop update` as the retry. Binaries already on the
-		// `entire` app just update in place.
-		if scoopAppName() == "cli" {
-			return `cmd.exe /D /C "scoop install entire/entire && scoop uninstall entire/cli && scoop reset entire"`
-		}
-		return "scoop update entire/entire"
 	}
-
-	if goos == goosWindows {
-		if releaseChannel(currentVersion) == installChannelNightly {
-			return windowsInstallNightlyCmd
-		}
-		return windowsInstallCmd
-	}
-
-	if releaseChannel(currentVersion) == installChannelNightly {
-		return "curl -fsSL https://entire.io/install.sh | bash -s -- --channel nightly"
-	}
-	return "curl -fsSL https://entire.io/install.sh | bash"
+	return fallbackInstallCommand(currentVersion)
 }
 
 // printNotification prints the version update notification to the user.
