+++ cmd/entire/cli/versioncheck/versioncheck_test.go
@@ -346,154 +346,23 @@ func TestParseGitHubRelease(t *testing.T) {
 // it without tripping goconst on repeated string literals.
 const brewUpgradeCmd = "brew upgrade --yes entire"
 
-// scoopExecutablePath is the pre-rename `cli` app dir; scoopEntireExecutablePath
-// is the renamed `entire` app dir. The Scoop update command is chosen by which
-// app dir the binary runs from, not by version.
-const scoopExecutablePath = `C:\Users\test\scoop\apps\cli\current\entire.exe`
-const scoopEntireExecutablePath = `C:\Users\test\scoop\apps\entire\current\entire.exe`
-
 // plainBinPath is a POSIX install under no recognized install manager.
 const plainBinPath = "/usr/local/bin/entire"
 
-// windowsLocalBinPath is a direct install.ps1 destination. windowsProgramFilesPath
-// is an unknown Windows path that matches none of the install-manager prefixes.
-const windowsLocalBinPath = `C:\Users\test\.local\bin\entire.exe`
-const windowsProgramFilesPath = `C:\Program Files\Entire\entire.exe`
-
-// TestScoopAppName pins the signal the rename migration keys off. The ""
-// results matter most: they are what keeps the `== "cli"` comparison in
-// UpdateCommandForCurrentBinary from misfiring on a non-Scoop binary that happens to live in a
-// directory named `cli`.
-func TestScoopAppName(t *testing.T) {
-	tests := []struct {
-		name     string
-		execPath func() (string, error)
-		want     string
-	}{
-		{
-			name:     "pre-rename cli app dir",
-			execPath: func() (string, error) { return scoopExecutablePath, nil },
-			want:     "cli",
-		},
-		{
-			name:     "renamed entire app dir",
-			execPath: func() (string, error) { return scoopEntireExecutablePath, nil },
-			want:     "entire",
-		},
-		{
-			name:     "non-scoop path in a cli directory",
-			execPath: func() (string, error) { return `C:\tools\cli\entire.exe`, nil },
-			want:     "",
-		},
-		{
-			name:     "posix install",
-			execPath: func() (string, error) { return plainBinPath, nil },
-			want:     "",
-		},
-		{
-			name:     "executable path unavailable",
-			execPath: func() (string, error) { return "", errors.New("not found") },
-			want:     "",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			original := executablePath
-			executablePath = tt.execPath
-			t.Cleanup(func() { executablePath = original })
-
-			if got := scoopAppName(); got != tt.want {
-				t.Errorf("scoopAppName() = %q, want %q", got, tt.want)
-			}
-		})
-	}
-}
+const miseExecutablePath = "/home/user/.local/share/mise/installs/entire/1.0.0/bin/entire"
 
 func TestUpdateCommandForCurrentBinary(t *testing.T) {
 	tests := []struct {
 		name           string
 		currentVersion string
-		goos           string
 		execPath       func() (string, error)
 		want           string
 	}{
 		{
-			name:           "homebrew stable cellar path uses brew command",
-			currentVersion: "1.0.0",
-			execPath:       func() (string, error) { return "/opt/homebrew/Cellar/entire/1.0.0/bin/entire", nil },
-			want:           brewUpgradeCmd,
-		},
-		{
-			name:           "homebrew stable cask path uses brew command",
-			currentVersion: "1.0.0",
-			execPath:       func() (string, error) { return "/opt/homebrew/bin/entire", nil },
-			want:           brewUpgradeCmd,
-		},
-		{
-			name:           "homebrew nightly path uses brew command",
-			currentVersion: "1.0.1-nightly.202604101200.abc1234",
-			execPath:       func() (string, error) { return "/opt/homebrew/bin/entire", nil },
-			want:           "brew upgrade --yes entire@nightly",
-		},
-		{
-			name:           "linuxbrew path",
-			currentVersion: "1.0.0",
-			execPath:       func() (string, error) { return "/home/linuxbrew/.linuxbrew/bin/entire", nil },
-			want:           brewUpgradeCmd,
-		},
-		{
 			name:           "mise path",
 			currentVersion: "1.0.0",
-			execPath:       func() (string, error) { return "/home/user/.local/share/mise/installs/entire/1.0.0/bin/entire", nil },
+			execPath:       func() (string, error) { return miseExecutablePath, nil },
 			want:           "mise upgrade entire",
-		},
-		{
-			name:           "scoop entire app updates in place",
-			currentVersion: "1.0.0",
-			execPath:       func() (string, error) { return scoopEntireExecutablePath, nil },
-			want:           "scoop update entire/entire",
-		},
-		{
-			name:           "scoop cli app routes through package rename regardless of version",
-			currentVersion: "1.0.0",
-			execPath:       func() (string, error) { return scoopExecutablePath, nil },
-			want:           `cmd.exe /D /C "scoop install entire/entire && scoop uninstall entire/cli && scoop reset entire"`,
-		},
-		{
-			name:           "unknown path stable falls back to stable curl command",
-			currentVersion: "1.0.0",
-			goos:           "linux",
-			execPath:       func() (string, error) { return plainBinPath, nil },
-			want:           "curl -fsSL https://entire.io/install.sh | bash",
-		},
-		{
-			name:           "unknown path nightly falls back to nightly curl command",
-			currentVersion: "1.0.1-nightly.202604101200.abc1234",
-			goos:           "linux",
-			execPath:       func() (string, error) { return plainBinPath, nil },
-			want:           "curl -fsSL https://entire.io/install.sh | bash -s -- --channel nightly",
-		},
-		{
-			name:           "executable error falls back to stable curl command",
-			currentVersion: "1.0.0",
-			goos:           "linux",
-			execPath:       func() (string, error) { return "", errors.New("not found") },
-			want:           "curl -fsSL https://entire.io/install.sh | bash",
-		},
-		{
-			name:           "windows unknown path stable uses install.ps1",
-			currentVersion: "1.0.0",
-			goos:           goosWindows,
-			execPath:       func() (string, error) { return windowsLocalBinPath, nil },
-			want:           windowsInstallCmd,
-		},
-		{
-			name:           "windows unknown path nightly uses install.ps1 -Channel nightly",
-			currentVersion: "1.0.1-nightly.202604101200.abc1234",
-			goos:           goosWindows,
-			execPath:       func() (string, error) { return windowsLocalBinPath, nil },
-			want:           windowsInstallNightlyCmd,
 		},
 	}
 
@@ -502,12 +371,6 @@ func TestUpdateCommandForCurrentBinary(t *testing.T) {
 			original := executablePath
 			executablePath = tt.execPath
 			t.Cleanup(func() { executablePath = original })
-
-			if tt.goos != "" {
-				origGOOS := goos
-				goos = tt.goos
-				t.Cleanup(func() { goos = origGOOS })
-			}
 
 			if got := UpdateCommandForCurrentBinary(tt.currentVersion); got != tt.want {
 				t.Errorf("UpdateCommandForCurrentBinary() = %q, want %q", got, tt.want)
